Move RaceHelper to src/game/client
