+++ CMakeLists.txt
@@ -624,8 +624,6 @@ if(CLIENT)
     components/particles.h
     components/players.cpp
     components/players.h
-    components/race.cpp
-    components/race.h
     components/race_demo.cpp
     components/race_demo.h
     components/scoreboard.cpp
@@ -644,6 +642,8 @@ if(CLIENT)
     gameclient.h
     lineinput.cpp
     lineinput.h
+    race.cpp
+    race.h
     render.cpp
     render.h
     render_map.cpp
